Related work items: #2029, #3884, #3914, #3917, #3935, #3938, #3939, #3942, #3943, #3944, #3947, #3950, #3952, #3954, #3960, #3961, #3962, #3963, #3964, #3965, #3966

+++ VerticalWarehouses/Common/Ferretto.VW.CommonUtils/Messages/Data/WeightAcquisitionCommandMessageData.cs
@@ -6,19 +6,17 @@ namespace Ferretto.VW.CommonUtils.Messages.Data
 {
     public class WeightAcquisitionCommandMessageData : IMessageData
     {
-
-
         #region Properties
 
         public CommandAction CommandAction { get; set; }
 
-        public decimal Displacement { get; set; }
+        public double Displacement { get; set; }
 
-        public decimal InitialPosition { get; set; }
+        public double InitialPosition { get; set; }
 
         public TimeSpan InPlaceSamplingDuration { get; set; }
 
-        public decimal NetWeight { get; set; }
+        public double NetWeight { get; set; }
 
         public MessageVerbosity Verbosity { get; set; }
